Move logic to merge items out of MergedSierraRecord
