+++ common/src/main/scala/uk/ac/wellcome/models/MergedSierraRecord.scala
@@ -26,39 +26,6 @@ case class MergedSierraRecord(
   version: Int = 0
 ) extends Transformable {
 
-  /** Given a new item record, construct the new merged row that we should
-    * insert into the merged database.
-    *
-    * Returns the merged record.
-    */
-  def mergeItemRecord(itemRecord: SierraItemRecord): MergedSierraRecord = {
-    if (!itemRecord.bibIds.contains(this.id)) {
-      throw new RuntimeException(
-        s"Non-matching bib id ${this.id} in item bib ${itemRecord.bibIds}")
-    }
-
-    // We can decide whether to insert the new data in two steps:
-    //
-    //  - Do we already have any data for this item?  If not, we definitely
-    //    need to merge this record.
-    //  - If we have existing data, is it newer or older than the update we've
-    //    just received?  If the existing data is older, we need to merge the
-    //    new record.
-    //
-    val isNewerData = this.itemData.get(itemRecord.id) match {
-      case Some(existing) =>
-        itemRecord.modifiedDate.isAfter(existing.modifiedDate)
-      case None => true
-    }
-
-    if (isNewerData) {
-      val itemData = this.itemData + (itemRecord.id -> itemRecord)
-      this.copy(itemData = itemData)
-    } else {
-      this
-    }
-  }
-
   def unlinkItemRecord(itemRecord: SierraItemRecord): MergedSierraRecord = {
     if (!itemRecord.unlinkedBibIds.contains(this.id)) {
       throw new RuntimeException(
